Refer to cluster by id, not name
If the cluster is replaced, but the node pool does not 'need to be',
the node-pool is not flagged for replacement when it should be.

It is deleted within GKE when the cluster itself is removed, but
terraform's state does not reflect this. When terraform then creates
the 'new' cluster, no node pool is created as, as far as terraform is
aware, it already exists.

By referring to the cluster by id which is not known at plan time for
a creation, the node pool is correctly flagged for replacement.

ref: https://github.com/hashicorp/terraform-provider-google/issues/
   11164#issuecomment-1050189855